requirements.txt
langchain
langchain-openai
langgraph
yfinance
python-dotenv